requirements.txt
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
easydict==1.13
filelock @ file:///croot/filelock_1700591183607/work
ftfy==6.2.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy @ file:///croot/sympy_1724938189289/work
threadpoolctl==3.5.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.5
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
